chore(refactor): moved skipcomments and filter ignored
