+++ packages/cli/src/services/processComparisonFile.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import { parseEnvFile } from './parseEnvFile.js';
-import { filterIgnoredKeys } from '../core/filterIgnoredKeys.js';
+import { filterIgnoredKeys } from '../core/helpers/filterIgnoredKeys.js';
 import { compareWithEnvFiles } from '../core/scan/compareScan.js';
 import { findDuplicateKeys } from '../core/duplicates.js';
 import { applyFixes } from '../core/fixEnv.js';
